Apply viewport effects to the mouse input coordinates
That way we can select objects even when the spherical mirror is on.  This is
actually a reintroduction of some old code that was removed.

+++ src/core/StelApp.hpp
@@ -240,7 +240,7 @@ private:
 	//! Handle mouse wheel.
 	void handleWheel(class QWheelEvent* event);
 	//! Handle mouse move.
-	void handleMove(int x, int y, Qt::MouseButtons b);
+	void handleMove(float x, float y, Qt::MouseButtons b);
 	//! Handle key press and release.
 	void handleKeys(class QKeyEvent* event);
 	//! Handle pinch on multi touch devices.
